+++ package/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - add `runtime` option
 - add `build.bundle` option
+- add `verbose` option
 
 ## 0.3.0 (2026-05-09)